pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mns-chatbot-backend"
version = "0.1.0"
description = "MNS Chatbot Backend"
dependencies = [
    "openai",
    "python-dotenv",
    "psycopg2-binary",
    "azure-functions",
    "pydantic"
]

[tool.pytest.ini_options]
pythonpath = ["."]